Fixes #2: fix docker builds
Docker layer caching broke the build. target/release/libcache.so that
was copied over from "builder" stage img was the result of the first
stage of compilation(dependencies only and no source code)

So the object file didn't contain anything that resembled this repo's source
code.

This patch employs a slightly modified caching technique(see Dockerfile)
which fixes the issue